Archetypes - Integration tests verifier approach
New approach for the execution of the ITs for the archetypes. It uses the
maven verifier component to generate the project from the archetype and
afterwards execute the required goal/s and check the results.

This approach works properly in windows in contrast to the default archetypes
ITs mechanism, which failed executing the runner script.