Archetypes - Blank messaging broker url
Add a non persistent localhost messaging broker url to avoid
ACS 6.1 enterprise to get stuck waiting for the ActiveMQ service
that it is not configured in the SDK generated projects.

A new documentation article will be created to show how to configure
the transformation service with docker for ACS 6.1 enterprise.