Archetypes - Improve mvn executable localisation
Change the run script to use M2_HOME environment variable if it is
set to locate the mvn executable. If it is not set, then the script
uses a plain mvn command. So if the M2_HOME env variable is not set,
then the mvn executable must be in the path.