Archetypes - Add build with IT supported task
Modify the run scripts to add the task: build_start_it_supported. After the
last changes regarding the way the project is compiled and retrieves the
IT dependencies, it losted the ability to run the project in docker with
IT support (including the libraries required to execute IT) but without
executing the tests. It only had build_start (no IT supported) and build_test
(IT supported but started and stopped the environment after IT execution).

This new task allows the execution of the docker environment with IT execution
supported to cover the next use cases:
- Running IT classes from IDE (for instance, to debug the execution).
- Running IT classes with hot reload enabled.